Move things to _jit_internal
